refactor: move snake gameboard file into its own folder and fix references
