+++ rn/src/components/snake/GameBoard.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useRef } from "react";
 import { View, StyleSheet, Animated, Easing } from "react-native";
-import { COLS } from "../game/snake/constants";
-import type { Snake } from "../game/snake/snake";
-import { palette } from "../theme";
-import { USE_NATIVE_DRIVER } from "../platform";
+import { COLS } from "../../game/snake/constants";
+import type { Snake } from "../../game/snake/snake";
+import { palette } from "../../theme";
+import { USE_NATIVE_DRIVER } from "../../platform";
 
 type Props = {
   snake: Snake;
